rag_pdf_api/requirements.txt
fastapi
uvicorn
PyMuPDF
sentence-transformers
faiss-cpu
openai
python-dotenv